This commit provides 3 notable changes:

- The weapon speed and skill delay timers have been integrated into a single action timer. The immediate effect of this is that players can no longer simultaneously use other skills while attacking mobs in melee. This was an unfair advantage to the melee skills before this change. However, it should be fairly easy to add something like dual-wield in future.

- The 'You can shoot again' messages have been removed, and a new timer has been added to the client UI instead to notify the player how long until the action timer has ran out. This warranted a protocol change, so the new version of the client and server are incompatible with previous revisions. With this change, it also becomes easy to block use skill commands to the server while the action timer is above zero. This would save a lot of bandwidth, and a little CPU time for the server.

- A bugfix to stop stats from wrapping around when too much food of poison is eaten at once. This is a little more than a bugfix though as some of the code is cleaned up around there and a little provision has been added so that food could possibly give buffs in future.

+++ trunk/server/src/include/sproto.h
@@ -221,8 +221,7 @@ void                        party_client_group_kill(object *member);
 void                        party_client_group_update(object *member, int flag);
 /* c_range.c */
 void                        fire(object *op, int dir);
-int                         fire_magic_tool(object *owner, object *op, int dir);
-int                         command_cast_spell(object *op, char *params);
+float						fire_magic_tool(object *owner, object *op, int dir);int                         command_cast_spell(object *op, char *params);
 /* c_wiz.c */
 int                         command_dm_connections(object *op, char *params);
 int                         command_dmload(object *op, char *params);
@@ -583,7 +582,7 @@ int                         attack_hth(object *pl, int dir, char *string);
 int                         skill_attack(object *tmp, object *pl, int dir, char *string);
 int                         do_skill_attack(object *tmp, object *op, char *string);
 int                         SK_level(object *op);
-void                        set_action_time(object *op, int ticks);
+void                        set_action_time(object *op, float ticks);
 int                         check_skill_action_time(object *op, object *skill);
 int                         get_skill_stat1(object *op);
 int                         get_skill_stat2(object *op);
